refactor(e2e): relocate prod-seq leftovers to single-node
